Adds `NativeElementReader`, lifts int size restriction
This PR:
* Adds `NativeElementReader` a pure-Rust implementation of the
`ElementReader` API. Iterates over a given byte slice and returns
materialized `OwnedElement` representations of the data.
* Modifies `tests/element_test_vectors.rs` to be able to run the
same `ion-tests` logic using both the `IonCElementReader` and the
new `NativeElementReader`. This includes the creation of a (very
long) skip list for the `NativeElementReader`.
* Renames `AnyInt` to `Integer` and moves it to the 'types' module
alongside `Decimal` and `Timestamp`.
* Renames `AnyUInt` to `UInteger` and moves it to the `types` module.
* Modifies the `UInt` decoder to be able to return `UInt`s of any
size, not just `u64`s.
* Modifies the `StreamReader` trait (and all implementations) to include
a `fn read_integer(&mut self) -> IonResult<Integer>` method that can
return an integer of any size.
* Fixes a bug in the `RawTextReader` that caused it to try and step into
container-typed `null` values.

+++ src/text/text_value.rs
@@ -1,5 +1,6 @@
 use crate::raw_symbol_token::RawSymbolToken;
 use crate::types::decimal::Decimal;
+use crate::types::integer::Integer;
 use crate::types::timestamp::Timestamp;
 use crate::IonType;
 
@@ -46,7 +47,7 @@ impl PartialEq<TextValue> for AnnotatedTextValue {
 pub(crate) enum TextValue {
     Null(IonType),
     Boolean(bool),
-    Integer(i64),
+    Integer(Integer),
     Float(f64),
     Decimal(Decimal),
     Timestamp(Timestamp),
